requirements.txt
streamlit==1.31.1
pandas==2.2.0
numpy==1.26.4
openai==1.12.0
bertopic==0.16.0
sentence-transformers==2.3.1
umap-learn==0.5.5
hdbscan==0.8.33
plotly==5.18.0
scikit-learn==1.4.0
scipy==1.12.0
matplotlib==3.8.2
seaborn==0.13.2
ipython==8.21.0
nbformat==5.9.2
nbconvert==7.14.2
python-dotenv==1.0.1
